Move comparison of state trajectories to md submodule.
