+++ src/msmhelper/msm/timescales.py
@@ -12,7 +12,7 @@ import decorit
 import numba
 import numpy as np
 
-from msmhelper.compare import _intersect as intersect
+from msmhelper.md.comparison import _intersect as intersect
 from msmhelper.msm.utils import linalg
 from msmhelper.statetraj import StateTraj
